updated setup script

+++ setup.php
@@ -68,8 +68,7 @@
   $options = getopt($short, $long);
 
   echo "ENTROPY installation and update script\n";
-  if (isset($options["h"]) or isset($options["help"]))
-  {
+  if (isset($options["h"]) or isset($options["help"])) {
     echo "\nUsage: setup.php [options]\n"
        . "Options:\n"
        . "  -u       --update       Update Entropy. Preserves configuration files and data. (default)\n"
@@ -82,8 +81,7 @@
 
   // invalid option combinations
   if (     (isset($options["u"]) or isset($options["update"])) 
-       and (isset($options["i"]) or isset($options["install"])) )
-  {
+       and (isset($options["i"]) or isset($options["install"])) ) {
     echo "\nERROR: contradicting options: clean install vs update only. Aborting...\n";
     exit(1);
   }
@@ -92,23 +90,18 @@
   if (isset($options["c"]))      $cfgfile = $options["c"];
   if (isset($options["config"])) $cfgfile = $options["config"];
 
-  if (isset($cfgfile))
-  {
+  if (isset($cfgfile)) {
     // read file
-    if (file_exists($cfgfile))
-    {
+    if (file_exists($cfgfile)) {
       $cfgfile = file_get_contents($cfgfile);
       $cfgfile = json_decode($cfgfile, true);
-    }
-    else
-    {
+    } else {
       echo "\nERROR: configuration file not found. Aborting...\n";
       exit(3);
     }
     
     // valid json?
-    if (empty($cfgfile))
-    {
+    if (empty($cfgfile)) {
       echo "\nERROR: invalid or empty configuration file. Aborting...\n";
       exit(4);
     }
@@ -119,14 +112,11 @@
   }
 
   // evaluate $cfg values
-  foreach ($cfg as $id => $value)
-  {
+  foreach ($cfg as $id => $value) {
     list($eval, $value) = cfgeval($id, $value);
 
-    if ($eval)
-      $cfg[$id] = $value; //corrected value
-    else
-    {
+    if ($eval) $cfg[$id] = $value; //corrected value
+    else {
       echo "\nERROR: configuration item " . $id . " not a valid " . gettype2($id) . "\n";
       exit(5);
     }
@@ -137,31 +127,22 @@
   elseif (isset($options["i"]) or isset($options["install"])) $cfg["forceinstall"] = True;
 
   // interactive mode
-  if ( !isset($options["d"]) and !isset($options["defaults"]) )
-  {    
-    foreach ($qinstall as $id => $value)
-    {
+  if ( !isset($options["d"]) and !isset($options["defaults"]) ) {    
+    foreach ($qinstall as $id => $value) {
       // if update/install options specified on command line, do'nt ask again
-      if ( !($id === "forceinstall" and (isset($options["u"]) or isset($options["update"]) or isset($options["i"]) or isset($options["install"]))) )
-      {
+      if ( !($id === "forceinstall" and (isset($options["u"]) or isset($options["update"]) or isset($options["i"]) or isset($options["install"]))) ) {
         echo "\n" . $value;
         echo "\nEnter value or accept default [" . (isset($cfg[$id]) ? $cfg[$id] : "") . "]: \n";
         
-        while (1)
-        {
+        while (1) {
           $response = trim(fgets(STDIN));
-          if (!empty($response))
-          {
+          if (!empty($response)) {
             list($eval, $response) = cfgeval($id, $response);
-            if ($eval)
-            {
+            if ($eval) {
               $cfg[$id] = trim($response);
               break;
-            }
-            else
-              echo "Invalid response; please answer with a " . gettype2($id) . ":\n";
-          }
-          else break;
+            } else echo "Invalid response; please answer with a " . gettype2($id) . ":\n";
+          } else break;
         }
       }
     }
@@ -174,13 +155,10 @@
 
 
   // determine if we do a clean install or an update
-  if (!file_exists($cfg["privpath"] . "entropy.conf.php") or $cfg["forceinstall"])
-  {
+  if (!file_exists($cfg["privpath"] . "entropy.conf.php") or $cfg["forceinstall"]) {
     $cleaninstall = True;
     echo "Clean installation...\n\n";
-  }
-  else
-  {
+  } else {
     $cleaninstall = False;
     echo "Update...\n\n";
   }
@@ -191,77 +169,62 @@
 
   // common tasks for both clean install and update
   echo "Copying files.\n\n";
-  if (!file_exists($cfg["privpath"])) 
-  {
+  if (!file_exists($cfg["privpath"])) {
     mkdir($cfg["privpath"], 0750, true);
     chgrp($cfg["privpath"], $cfg["htgroup"]);
   }
-  if (!file_exists($cfg["pubpath"]))
-  {
+  if (!file_exists($cfg["pubpath"])) {
     mkdir($cfg["pubpath"], 0750, true);
     chgrp($cfg["pubpath"], $cfg["htgroup"]);
   }
 
-  rcopy("./inc/",                  $cfg["privpath"] . "inc/",      $cfg["htgroup"]);
-  rcopy("./LICENSE",               $cfg["privpath"] . "LICENSE",   $cfg["htgroup"]);
-  rcopy("./README.md",             $cfg["privpath"] . "README.md", $cfg["htgroup"]);
-  
-  rcopy("./public_html/css/",      $cfg["pubpath"] . "css/",       $cfg["htgroup"]);
-  rcopy("./public_html/img/",      $cfg["pubpath"] . "img/",       $cfg["htgroup"]);
-  rcopy("./public_html/js/",       $cfg["pubpath"] . "js/",        $cfg["htgroup"]);
-  rcopy("./public_html/auth.php",  $cfg["pubpath"] . "auth.php",   $cfg["htgroup"]);
-  rcopy("./public_html/tools.php", $cfg["pubpath"] . "tools.php",  $cfg["htgroup"]);
-  rcopy("./public_html/index.php", $cfg["pubpath"] . "index.php",  $cfg["htgroup"]);
+  // recursively copy all subdirectories in ./app and ./public_html
+  foreach (scandir2("./app/") as $f) {
+    rcopy("./app/" . $f, $cfg["privpath"] . $f, $cfg["htgroup"]);
+  }
+  foreach (scandir2("./public_html/") as $f) {
+    rcopy("./public_html/" . $f, $cfg["pubpath"] . $f, $cfg["htgroup"]);
+  }
 
   // clean install only
-  if ($cleaninstall)
-  {
+  if ($cleaninstall) {
     // copy data folder (writable for htgroup)
     rcopy("./data/", $cfg["privpath"] . "data/", $cfg["htgroup"], True);
 
     // interactive mode: ask config file questions
-    if ( !isset($options["d"]) and !isset($options["defaults"]) )
-    {    
-      foreach ($qconfig as $id => $value)
-      {
+    if ( !isset($options["d"]) and !isset($options["defaults"]) ) {    
+      foreach ($qconfig as $id => $value) {
         echo "\n" . $value;
         echo "\nEnter value or accept default [" . (isset($cfg[$id]) ? $cfg[$id] : "") . "]: \n";
 
-        while (1)
-        {
+        while (1) {
           $response = trim(fgets(STDIN));
-          if (!empty($response))
-          {
+          if (!empty($response)) {
             list($eval, $response) = cfgeval($id, $response);
-            if ($eval)
-            {
+            if ($eval) {
               $cfg[$id] = trim($response);
               break;
-            }
-            else
-              echo "Invalid response; please answer with a " . gettype2($id) . ":\n";
-          }
-          else break;
+            } else echo "Invalid response; please answer with a " . gettype2($id) . ":\n";
+          } else break;
         }
       }
     }
 
     // build entropy.conf.php
     echo "Build configuration file: entropy.conf.php\n";
-    $in = fopen("./entropy.conf.php", "r");
+    $in = fopen("./default_config/entropy.conf.php", "r");
     $out = fopen($cfg["privpath"] . "entropy.conf.php", "w");
     mkconf($in, $out, $cfg);
 
     // build install.conf.php
     echo "Build configuration file: install.conf.php\n";
-    $in = fopen("./public_html/install.conf.php", "r");
+    $in = fopen("./default_config/install.conf.php", "r");
     $out = fopen($cfg["pubpath"] . "install.conf.php", "w");
     mkconf($in, $out, $cfg);
 
     // build users.json
     echo "Build configuration file: users.json\n";
-    if (!isset($cfg["users_file"]) and !isset($cfg["config_path"]))
-    { 
+    if (!isset($cfg["users_file"]) and !isset($cfg["config_path"])) { 
       $usercfg = Array( $cfg["admin_login"] => Array( "name"        => $cfg["admin_name"],
                                                       "institution" => (isset($cfg["admin_inst"])  ? $cfg["admin_inst"]  : ""),
                                                       "email"       => (isset($cfg["admin_email"]) ? $cfg["admin_email"] : ""),
@@ -276,9 +239,7 @@
                       );
       $usercfg = json_encode($usercfg, JSON_PRETTY_PRINT);
       file_put_contents($cfg["privpath"] . "data/config/users.json" , $usercfg);
-    }
-    else
-      echo "WARNING: cannot make the users.json file if the USERS_FILE or CONFIG_PATH constants were changed from default values!";
+    } else echo "WARNING: cannot make the users.json file if the USERS_FILE or CONFIG_PATH constants were changed from default values!";
   }
 
   echo "\nFinished.\n\n";
@@ -291,29 +252,35 @@
   //   INTERNAL FUNCTIONS   //
   //------------------------//
 
+  
+  // array of files and subdirs (without . and ..) for a given dir
+  function scandir2($dir)
+  {
+    if (is_dir($dir)) {
+      $files = scandir($dir);
+      return array_diff($files, [".", ".."]);
+    }
+    return array();
+  }
+  
   // removes files and non-empty directories
   function rrmdir($dir)
   {
-    if (is_dir($dir))
-    {
-      $files = scandir($dir);
-      foreach ($files as $file)
-        if ($file != "." && $file != "..")
-          rrmdir("$dir/$file");
+    if (is_dir($dir)) {
+      $files = scandir2($dir);
+      foreach ($files as $file) rrmdir("$dir/$file");
       rmdir($dir);
+    } else {
+      if (file_exists($dir)) unlink($dir);
     }
-    else
-      if (file_exists($dir))
-        unlink($dir);
   }
 
   // copies files and non-empty directories
   function rcopy($src, $dst, $group, $writable = False)
   {
-    if (file_exists($dst))
-      rrmdir($dst);
-    if (is_dir($src))
-    {
+     if (file_exists($dst)) rrmdir($dst);
+
+    if (is_dir($src)) {
       mkdir($dst);
       chmod($dst, ($writable ? 0770 : 0750));        
       chgrp($dst, $group);
@@ -321,34 +288,29 @@
       foreach ($files as $file)
         if ($file != "." && $file != "..")
           rcopy("$src/$file", "$dst/$file", $group, $writable);
-    }
-    else
+    } else {
       if (file_exists($src))
       {
         copy($src, $dst);
         chmod($dst, ($writable ? 0660 : 0640));
         chgrp($dst, $group);
       }
+    }
   }
 
   function mkconf($in, $out, $cfg)
   {
-    if ($in) 
-    {
-      while (($line = fgets($in)) !== false) 
-      {
+    if ($in) {
+      while (($line = fgets($in)) !== false) {
         $lineparts = explode("//", trim($line), 2);  // only consider the part of the line before eventual comments
 
-        if (substr(strtolower($lineparts[0]), 0, 5) == 'const')
-        {
+        if (substr(strtolower($lineparts[0]), 0, 5) == 'const') {
           $static = explode('=', $lineparts[0], 2)[0];  // only consider the part before "="
           $static = trim(str_replace("CONST", "", strtoupper($static)));
           
-          foreach ($cfg as $item => $value)
-          {
-            if ($static == strtoupper($item))   // found cfg-item in line -> replace it
-            {
-               $lineparts[0] = '  const ' . strtoupper($item) . ' = ' . putvalue($item, $value) . '; ';
+          foreach ($cfg as $item => $value) {
+            if ($static == strtoupper($item)) {  // found cfg-item in line -> replace it
+              $lineparts[0] = '  const ' . strtoupper($item) . ' = ' . putvalue($item, $value) . '; ';
               $line = implode(" //", $lineparts) . "\n";
               echo "  - set " . strtoupper($item) . " -> " . $value . "\n";
             }
@@ -367,8 +329,7 @@
 
   function cfgeval($item, $value)
   {
-    switch (gettype2($item))
-    {
+    switch (gettype2($item)) {
       case "boolean":
         $value = filter_var($value, FILTER_VALIDATE_BOOLEAN);
         $eval = True;  //True, "true", "on", "yes", "1" are True, all others are False, thus evaluation always succeeds by design
@@ -409,8 +370,7 @@
 
   function putvalue($item, $value)
   {
-    switch (gettype2($item))
-    {
+    switch (gettype2($item)) {
       case "boolean":
         return ($value ? "True" : "False");
         break;
